requirements.txt
huggingface_hub==0.25.2
yfinance
transformers
pandas
pandas-ta
ta
textblob
aiohttp
python-dotenv
newsapi-python
gradio
torch
groq